requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
flask-bcrypt>=1.0.1
mysql-connector-python>=8.1.0
PyJWT>=2.8.0
pytest>=7.4.0

